Support fallback error image URL
